componetize filtered reducers
